pyproject.toml
[tool.poetry]
name = "knewkarma"
version = "8.0.0"
description = "Reddit-data analysis toolkit"
authors = ["Ritchie Mwewa <rly0nheart@duck.com>", "Knew Karma IO <redacted@redacted.com>"]
readme = "README.md"
license = "GPL-3.0+"
keywords = ["reddit", "data-analysis"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Natural Language :: English"
]

[tool.poetry.urls]
sponsor = "https://opencollective.com/knewkarma"
homepage = "https://knewkarma.readthedocs.io"
issues = "https://github.com/knewkarma-io/knewkarma/issues"
repository = "https://github.com/knewkarma-io/knewkarma"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pandas = "^2.1.4"
praw = "^7.7.1"
pydantic = "^2.11.5"
rich-click = "^1.8.9"

[tool.poetry.group.dev.dependencies]
myst-parser = "^3.0.1"
poetry = "^1.8.3"
sphinx = "^7.4.7"
sphinx-rtd-theme = "^2.0.0"

[tool.poetry.scripts]
rauth = "knewkarma.cli.commands:auth"
ruser = "knewkarma.cli.commands:user"
rusers = "knewkarma.cli.commands:cmd_users"
rsubreddit = "knewkarma.cli.commands:subreddit"
rsubreddits = "knewkarma.cli.commands:cmd_subreddits"
rsearch = "knewkarma.cli.commands:cmd_search"
rpost = "knewkarma.cli.commands:post"
rposts = "knewkarma.cli.commands:cmd_posts"
knewkarma = "knewkarma.cli.commands:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
